+++ packages/angular-table/tests/test-setup.ts
@@ -1,4 +1,5 @@
 import '@analogjs/vite-plugin-angular/setup-vitest'
+import '@testing-library/jest-dom/vitest'
 
 import {
   BrowserDynamicTestingModule,
