+++ packages/match-sorter-utils/tests/match-sorter-utils.test.ts
@@ -1,3 +1,4 @@
+import { describe, expect, it } from 'vitest'
 import { rankings, rankItem } from '../src'
 
 interface Person {
