+++ packages/react-table/tests/features/RowSelection.test.tsx
@@ -1,3 +1,4 @@
+import { expect, test } from 'vitest'
 import {
   ColumnDef,
   flexRender,
