+++ packages/table-core/tests/RowSelection.test.ts
@@ -1,3 +1,4 @@
+import { describe, expect, it } from 'vitest'
 import {
   ColumnDef,
   createColumnHelper,
