+++ packages/table-core/vite.config.ts
@@ -1,11 +1,13 @@
 import { defineConfig } from 'vitest/config'
+import packageJson from './package.json'
 
 export default defineConfig({
   test: {
+    name: packageJson.name,
+    dir: './tests',
     watch: false,
-    setupFiles: ['test-setup.ts'],
     environment: 'jsdom',
+    setupFiles: ['./tests/test-setup.ts'],
     globals: true,
-    dir: '__tests__',
   },
 })
